+++ src/Controller/SecurityController.php
@@ -9,18 +9,55 @@ use Symfony\Component\Security\Http\Authentication\AuthenticationUtils;
 
 class SecurityController extends AbstractController
 {
+    public function __construct(
+        private readonly array $options,
+    ) {
+    }
+
     #[Route(path: '/login', name: 'app_login')]
     public function login(AuthenticationUtils $authenticationUtils): Response
     {
-        // get the login error if there is one
+        // https://symfony.com/bundles/EasyAdminBundle/current/dashboards.html#login-form-template
         $error = $authenticationUtils->getLastAuthenticationError();
-
-        // last username entered by the user
         $lastUsername = $authenticationUtils->getLastUsername();
 
-        return $this->render('security/login.html.twig', [
-            'last_username' => $lastUsername,
+        return $this->render('@EasyAdmin/page/login.html.twig', [
+            // parameters usually defined in Symfony login forms
             'error' => $error,
+            'last_username' => $lastUsername,
+
+            // the translation_domain to use (define this option only if you are
+            // rendering the login template in a regular Symfony controller; when
+            // rendering it from an EasyAdmin Dashboard this is automatically set to
+            // the same domain as the rest of the Dashboard)
+            'translation_domain' => 'admin',
+
+            // by default EasyAdmin displays a black square as its default favicon;
+            // use this method to display a custom favicon: the given path is passed
+            // "as is" to the Twig asset() function:
+            // <link rel="shortcut icon" href="{{ asset('...') }}">
+            'favicon_path' => '/favicon.svg',
+
+            // the title visible above the login form (define this option only if you are
+            // rendering the login template in a regular Symfony controller; when rendering
+            // it from an EasyAdmin Dashboard this is automatically set as the Dashboard title)
+            'page_title' => $this->options['page_title'] ?? 'RPA Process Overview',
+
+            // the string used to generate the CSRF token. If you don't define
+            // this parameter, the login form won't include a CSRF token
+            'csrf_token_intention' => 'authenticate',
+
+            // the URL users are redirected to after the login (default: '/admin')
+            'target_path' => $this->generateUrl('app_default'),
+
+            // the label displayed for the username form field (the |trans filter is applied to it)
+            'username_label' => 'Email',
+
+            // the label displayed for the password form field (the |trans filter is applied to it)
+            'password_label' => 'Password',
+
+            // the label displayed for the Sign In form button (the |trans filter is applied to it)
+            'sign_in_label' => 'Log in',
         ]);
     }